requirements.txt
python-dotenv==1.2.2
pypdf==6.9.2
sentence-transformers==5.3.0
qdrant-client==1.17.1
google-generativeai==0.8.6
elasticsearch==8.19.3
